requirements.txt
streamlit>=1.28.0
requests>=2.31.0
toml>=0.10.2
pywebview>=4.4.1
pyinstaller>=6.3
pytest>=8.3.0
pytest-cov>=5.0.0
pytest-xdist>=3.6.0
google-generativeai>=0.8.3
